flake: mkOutputs: packages: fix relative paths in external presentations
The root Asciidoctor attribute for external presentations causes build
failures because it refers to the directory containing the evaluated
file, instead of the working directory.

This change reverts the root attribute to a Nix store path, aligning
with commit 9c354fd89b6c ("flake: mkOutputs: packages: do not set root
to temporary directory"), and replaces it with '.' afterwards.

Fixes: 716858e129e7 ("flake: mkOutputs: packages: increase external presentation deployability")